[refactor]: 성적조회 컴포넌트 타입 정의 파일 생성 및 컴포넌트별 import (#14)
